Ai-Muffin is now a Hatchery builder
the bot is now a very enthusiastic hatchery builder. it has a32*32 combination of co-ords of where it could build a hatch, and god dammit is it gonna try all of them before anything else.

training will take longer but the theory is in the long term it will recognise safe and useful expansion.